Cloud-Run/secfillingspublisher/requirements.txt
Flask>=2.0
requests>=2.28
google-cloud-bigquery>=3.0.0




